CF-1.8 compliance at write time in arctic_reconst pipeline
- config_utils: rewrite build_var_encoding around explicit PACKING_SPEC
  (14 vars, fixed int16 ranges, _FillValue=-32768); add print_packing_spec;
  set TIME_ENCODING _FillValue=None; build_global_attrs now writes
  Conventions='CF-1.8', expands short license tags, populates history.
- E_finalize_with_glorys: ADH standard_name -> geopotential_height; assign
  latitude/longitude as non-dim coords so xarray auto-emits
  coordinates='latitude longitude' per CF 5.6; coerce ocean_mask
  flag_values to int8 to match variable dtype (CF 3.5); timestamp daily
  files at 12:00 UTC (matches published archive convention); call
  print_packing_spec once per run for reproducibility.
- A_create_ocean_mask: ocean_mask.flag_values dtype int8 (was uint8);
  elevation and latitude/longitude stored as float64 (were float32).
- glorys_regrid, geos_currents, D_arctic_reconstruction: compute in
  float64 throughout (only quantize at the final write step).

Acceptance test on arctic_25km 2011-01-04: compliance-checker -t cf:1.8
passes; numerical diffs vs the published archive match the predicted
int16 packing precision for every changed variable.

+++ 4_arctic_reconst_ease/A_create_ocean_mask.py
@@ -184,7 +184,7 @@ def interpolate_gebco(gebco_path, x_ease, y_ease):
 
     tgt_x2d, tgt_y2d = np.meshgrid(x_ease, y_ease)
     pts = np.column_stack([tgt_y2d.ravel(), tgt_x2d.ravel()])
-    elev = interp(pts).reshape(len(y_ease), len(x_ease)).astype(np.float32)
+    elev = interp(pts).reshape(len(y_ease), len(x_ease))
     print(f"Bathymetry interpolation done — "
           f"range [{np.nanmin(elev):.0f}, {np.nanmax(elev):.0f}] m")
     return elev
@@ -251,7 +251,7 @@ def create_static_dataset(cfg):
             'ocean_mask': (['y_ease', 'x_ease'], ocean_mask, {
                 'standard_name': 'sea_binary_mask',
                 'long_name': 'Ocean mask',
-                'flag_values': np.array([0, 1], dtype=np.uint8),
+                'flag_values': np.array([0, 1], dtype=np.int8),
                 'flag_meanings': 'land ocean',
                 'grid_mapping': 'ease_grid_mapping',
                 **({'comment': 'Restricted to the configured lat/lon domain '
@@ -266,13 +266,13 @@ def create_static_dataset(cfg):
                 'source': f'Interpolated from 1 km GEBCO to {label}',
                 'grid_mapping': 'ease_grid_mapping',
             }),
-            'latitude': (['y_ease', 'x_ease'], lat_2d.astype(np.float32), {
+            'latitude': (['y_ease', 'x_ease'], lat_2d.astype(np.float64), {
                 'standard_name': 'latitude',
                 'long_name': 'Latitude',
                 'units': 'degrees_north',
                 'grid_mapping': 'ease_grid_mapping',
             }),
-            'longitude': (['y_ease', 'x_ease'], lon_2d.astype(np.float32), {
+            'longitude': (['y_ease', 'x_ease'], lon_2d.astype(np.float64), {
                 'standard_name': 'longitude',
                 'long_name': 'Longitude',
                 'units': 'degrees_east',
